New commands and minor fixes
New commands:
music:
Still being worked on, issues with user information not being updated causing channel.join() to fail as it never recieves a return. Probably an issue with the async/on.message.

Panel and PanelAPICheck:
These are admin commands that are still being worked on, and will be used for remotely and easily managing certain features of our panel via its Web (curl based) Api.
Purge:
Purge command for bulk deleting text. Options: Specific @user or all, a channel, and a number of messages to delete.

Fixes:

stickymessage and Unstick:
Updated Category to make it more clean and better represent the required permissions for the command.

UPNPCFix:
updated permissions required and category, as it should not be run by 'anyone'.

Index:
Commented out bot exit on error. This was useful for testing, but should not be in production versions or anyone can shut the bot down by accident...

Package:
Added more libraries/dependencies for use with future commands like music.